Find orphan concepts for all concept sets in a cohort

Searches for concepts that should belong to the concept sets in a cohort definition but don't, for example because of missing source-to-standard concept maps, or erroneous hierarchical relationships.

findCohortOrphanConcepts(
  connectionDetails = NULL,
  connection = NULL,
  cdmDatabaseSchema,
  oracleTempSchema = NULL,
  baseUrl = NULL,
  webApiCohortId = NULL,
  cohortJson = NULL,
  conceptCountsDatabaseSchema = cdmDatabaseSchema,
  conceptCountsTable = "concept_counts"
)

Arguments

connectionDetails
An object of type connectionDetails as created using the createConnectionDetails function in the DatabaseConnector package. Can be left NULL if connection is provided.

connection
An object of type connection as created using the connect function in the DatabaseConnector package. Can be left NULL if connectionDetails is provided, in which case a new connection will be opened at the start of the function, and closed when the function finishes.

cdmDatabaseSchema
Schema name where your patient-level data in OMOP CDM format resides. Note that for SQL Server, this should include both the database and schema name, for example 'cdm_data.dbo'.

oracleTempSchema
Should be used in Oracle to specify a schema where the user has write privileges for storing temporary tables.

baseUrl
The base URL for the WebApi instance, for example: "http://server.org:80/WebAPI". Needn't be provided if cohortJson is provided.

webApiCohortId
The ID of the cohort in the WebAPI instance. Needn't be provided if cohortJson is provided.

cohortJson
A character string containing the JSON of a cohort definition. Needn't be provided if baseUrl and webApiCohortId are provided.

conceptCountsDatabaseSchema
Schema name where your concept counts table resides. Note that for SQL Server, this should include both the database and schema name, for example 'scratch.dbo'.

conceptCountsTable
Name of the concept counts table. This table can be created using the createConceptCountsTable.

Value

A data frame with orphan concepts, with counts how often the code was encountered in the CDM.

Details

Logically, this function performs the following steps for each concept set expression in the cohort definition:

Given the concept set expression, find all included concepts.

Find all names of the input concepts, including synonyms, and the names of source concepts that map to them.

Search for concepts (standard and source) that contain any of those names as substring.

Filter those concepts to those that are not in the original set of concepts (i.e. orphans).

Restrict the set of orphan concepts to those that appear in the CDM database and across network concept prevalence (as either source concept or standard concept).